Change the way from and to block number arguments work.
First of all, calling pg_check() without range arguments always threw an
error if the table was completely empty, with zero blocks.

Change the block number arguments to 64-bit ints, because that way it's
easier to pass block numbers larger than 2^31 (otherwise you'd need to pass
them in as negative int4s, that then get casted internally to unsigned ints).

Also change the end block number arguments to be exclusive, not inclusive.
So e.g. if you specify range 5-8, it scans pages 5, 6 and 7, but not 8.
That generally makes it easier to write programs that call pg_check in a
loop with increasing page numbers, and also simplifies the logic in
pg_check itself a little bit, IMHO.