Handle when a tuple has fewer attributes than in the relation.
Postgres allows this situation to occur, and simply assumes that the trailing
attributes are NULL. This change makes pg_check only examine up to the number
of attributes actually in the tuple.

I don't know how to create this situation in a test, I've only seen it in
production data so far.