Limit checks in check_page_header to PG_PAGE_LAYOUT_VERSION
Some of the checks comparing offsets on the page only work with the
current page format (PG_PAGE_LAYOUT_VERSION), so skip them when
dealing with an older page format. We treat it as error, so that
we don't try doing any additional checks on the page (parse the
items etc.) as it's likely to fail.

This shouldn't be a big deal, as v. 4 is used since PostgreSQL 8.3.