Make index cross-check compatible with page range
Cross-checking table with indexes was originally supported only when
the whole table was scanned. But it's fairly trivial to support that
by simply building ignoring index items pointing to different pages.
We still need to scan the whole index, of course.

There used to be two pg_check_table() functions, one for cross-check
and one for range check. Now we only have one supporting both types
of functionality.